+++ GameObject.h
@@ -1,9 +1,6 @@
-#include <GL/glut.h>
+#pragma once
 
-struct GameObject;
-class Ball;
-class Paddle;
-class Boundary;
+#include <GL/glut.h>
 
 /**
  * @brief A struct that represents a game object. (everything on the screen)
@@ -37,7 +34,7 @@ public:
 
     GameObject();
     GameObject(float xPos, float yPos, float xLen, float yLen, float xVel, float yVel);
-    ~GameObject();
+    virtual ~GameObject();
 
     /**
      * @brief Function to draw the object on the screen.
@@ -65,59 +62,4 @@ public:
      * @return false if not colliding
      */
     bool isColliding(GameObject other);
-};
-
-/**
- * @brief Ball Game Object. Has collision and is a core component of the game
- *
- */
-class Ball : public GameObject
-{
-public:
-    typedef GameObject super;
-    float rad;
-
-    Ball();
-
-    Ball(float x, float y, float r, float xv, float yv);
-
-    void update(Paddle p1, Paddle p2);
-    void draw();
-    void reset();
-
-private:
-    float initialXPos;
-    float initialYPos;
-};
-
-/**
- * @brief Paddle Game Object. Has collision and is the playable part of the game.
- *
- */
-class Paddle : public GameObject
-{
-public:
-    typedef GameObject super;
-    float speed = 0.02f;
-
-    Paddle();
-    Paddle(float x, float y, float xl, float yl, float xv, float yv);
-
-    void draw();
-    void update();
-};
-
-/**
- * @brief Boundary Game Object. Has collision and acts as a wall for both the ball and the paddles.
- *
- */
-class Boundary : public GameObject
-{
-public:
-    typedef GameObject super;
-    
-    Boundary();
-    Boundary(float x, float y, float xl, float yl, float xv, float yv);
-
-    void draw();
 };
